+++ spec/support/mock_draw_system.rb
@@ -1,9 +1,9 @@
-class MockDrawSystem < Baku::System
+class MockDrawSystem < Baku::ComponentSystem
   def initialize
     super([MockComponent], :draw)
   end
 
-  def process_entity(entity, test)
-    test.count += 1
+  def process_entity(entity, mock)
+    mock.count += 1
   end
 end