+++ spec/support/mock_update_system.rb
@@ -1,9 +1,9 @@
-class MockUpdateSystem < Baku::System
+class MockUpdateSystem < Baku::ComponentSystem
   def initialize
     super([MockComponent], :update)
   end
 
-  def process_entity(entity, test)
-    test.count += 1
+  def process_entity(entity, mock)
+    mock.count += 1
   end
 end